+++ application-shell/package.json
@@ -9,12 +9,18 @@
     "start:live": "webpack serve --open --mode development --live-reload --hot"
   },
   "license": "MIT",
-  "author": "",
+  "author": {
+    "name": "Jack Herrington",
+    "email": "jherr@pobox.com"
+  },
   "devDependencies": {
     "@babel/core": "^7.15.8",
     "@babel/plugin-transform-runtime": "^7.15.8",
     "@babel/preset-env": "^7.15.8",
     "@babel/preset-react": "^7.14.5",
+    "@babel/preset-typescript": "^7.10.4",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.2",
     "autoprefixer": "^10.1.0",
     "babel-loader": "^8.2.2",
     "css-loader": "^6.3.0",
@@ -22,16 +28,13 @@
     "postcss": "^8.2.1",
     "postcss-loader": "^4.1.0",
     "style-loader": "^3.3.0",
+    "typescript": "^4.5.2",
     "webpack": "^5.57.1",
     "webpack-cli": "^4.9.0",
     "webpack-dev-server": "^4.3.1"
   },
   "dependencies": {
-    "@babel/runtime": "^7.13.10",
-    "bootstrap": "^5.1.3",
-    "dotenv-webpack": "^7.1.0",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.4.0",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.3.0"
   }
