+++ application-shell/src/App.tsx
@@ -1,11 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import AppRoutes from "./AppRoutes";
-import "bootstrap/dist/css/bootstrap.min.css";
 
-const App = () => (
-  <div>
-    <AppRoutes />
-  </div>
-);
+const App = () => <AppRoutes />;
 ReactDOM.render(<App />, document.querySelector("#app"));
