+++ application-shell/src/AppRoutes.tsx
@@ -1,15 +1,16 @@
 import React, { lazy, Suspense } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { PageErrorBounday } from "./errorboundaries/PageErrorBoundary";
+import Loading from "./loading/Loading";
 const Login = lazy(() => import("authentication/Login"));
 const Register = lazy(() => import("authentication/Register"));
 const HomePageLayout = lazy(() => import("./HomePageLayout"));
-const Home = lazy(() => import("application/Home"));
+const Home = lazy(() => import("component/Home"));
 const Users = lazy(() => import("authentication/Users"));
 const UserDetails = lazy(() => import("authentication/UserDetails"));
 export default function AppRoutes() {
   return (
-    <Suspense fallback={"Loading"}>
+    <Suspense fallback={<Loading />}>
       <Router>
         <Routes>
           <Route
