--- application-shell/src/components/home/Home.jsx
@@ -1,16 +1,0 @@
-import React, { useEffect, useState } from 'react'
-import { getHomeData } from '../../../../authentication-shell/src/services/service'
-
-export default function () {
-    const [data, setData] = useState();
-    useEffect(() => {
-        getHomeData().then((res) => { setData(res.data) })
-    }, [])
-
-    return (
-        <div>
-            <h1>Home</h1>
-            {data&&<h1>{data}</h1>}
-        </div>
-    )
-}
