--- application-shell/src/errorboundaries/ComponentErrorBoundary.jsx
@@ -1,45 +1,0 @@
-import React from "react";
-export const AsyncLoader = ({ children }) => {
-  return (
-    <ErrorBoundary>
-      <React.Suspense
-        fallback={
-          <div className="text-center">
-            <div className="spinner-border text-danger" role="status">
-              <span className="visually-hidden">Loading...</span>
-            </div>
-          </div>
-        }
-      >
-        {children}
-      </React.Suspense>
-    </ErrorBoundary>
-  );
-};
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError(error) {
-    if (error) return { hasError: true };
-  }
-
-  componentDidCatch(error, errorInfo) {
-    console.error(error, errorInfo);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <div className='text-center'>
-          <div className="spinner-border text-danger" role="status">
-            <span className="visually-hidden">Loading...</span>
-          </div>
-        </div>);
-    }
-
-    return this.props.children;
-  }
-}
