+++ application-shell/src/index.ts
