+++ application-shell/src/loading/loading.css
