+++ authentication-shell/.babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": ["@babel/preset-react", "@babel/preset-env"],
+  "presets": ["@babel/preset-typescript", "@babel/preset-react", "@babel/preset-env"],
   "plugins": [
      ["@babel/transform-runtime"]
   ]
