--- authentication-shell/src/App.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-
-const App = () => (
-    <>
-        <div className="container">
-            <div>Name: authentication-shell</div>
-            <div>Framework: react</div>
-            <div>Language: JavaScript</div>
-            <div>CSS: Empty CSS</div>
-        </div>
-    </>
-);
-ReactDOM.render(<App />, document.querySelector("#app"));
