+++ authentication-shell/src/index.ts
