+++ authentication-shell/src/pages/register/Register.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Form, Button } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "../../index.css";
+import "./register.css";
 import { useNavigate } from "react-router-dom";
 import {
   regexForName,
