+++ authentication-shell/src/pages/user-details/UserDetails.tsx
@@ -3,7 +3,11 @@ import { useParams } from "react-router-dom";
 import { getUserDetails } from "../../services/service";
 export default function UserDetails() {
   const { id } = useParams();
-  const [user, setUser] = useState();
+  const [user, setUser] = useState({
+    firstName: "",
+    lastName: "",
+    isActive: true,
+  });
   useEffect(() => {
     console.log(id);
     getUserDetails(id).then((res) => {
