+++ authentication-shell/src/pages/users/users.css
@@ -8,3 +8,44 @@
 .customCard img {
     height: 200px;
 }
+
+.forgetPassword {
+    font-size: 12px;
+    letter-spacing: 2px;
+    color: #e1e1e1;
+    font-family: inherit;
+    cursor: pointer;
+    position: relative;
+    border: none;
+    background: none;
+    text-transform: uppercase;
+    transition-timing-function: cubic-bezier(0.25, 0.8, 0.25, 1);
+    transition-duration: 400ms;
+    transition-property: color;
+}
+
+
+.forgetPassword:focus,
+.forgetPassword:hover {
+    color: #fff;
+}
+
+.forgetPassword:focus:after,
+.forgetPassword:hover:after {
+    width: 100%;
+    left: 0%;
+}
+
+.forgetPassword:after {
+    content: "";
+    pointer-events: none;
+    bottom: -2px;
+    left: 50%;
+    position: absolute;
+    width: 0%;
+    height: 2px;
+    background-color: #fff;
+    transition-timing-function: cubic-bezier(0.25, 0.8, 0.25, 1);
+    transition-duration: 400ms;
+    transition-property: width, left;
+}
