+++ authentication-shell/src/resources/constants.ts
@@ -1,0 +1,5 @@
+export const regexForEmail =
+  /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/;
+export const regexForPassword =
+  /^((?=\S*?[A-Z])(?=\S*?[a-z])(?=\S*?\d).{6,})\S$/; // Checks that a password has a minimum of 6 characters, at least 1 uppercase letter, 1 lowercase letter, and 1 number with no spaces.
+export const regexForName = /^[A-Za-z]{2,15}$/;
