--- authentication-shell/src/utils/error.js
@@ -1,9 +1,0 @@
-const error = (object) => {
-  let message = "Undefined Error";
-  if (object && object.message) {
-    message = object.message;
-  }
-  return message;
-};
-
-export default error;
