+++ component-shell/.babelrc
