+++ component-shell/.gitignore
