+++ component-shell/Dockerfile
@@ -10,6 +10,6 @@ COPY . .
 
 RUN npm run build
 
-EXPOSE 3001
+EXPOSE 3002
 
 CMD [ "npm", "run", "start" ]
