+++ component-shell/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "authentication-shell",
+  "name": "component-shell",
   "version": "1.0.0",
   "scripts": {
     "build": "webpack --mode production",
     "build:dev": "webpack --mode development",
-    "build:start": "cd dist && PORT=3001 npx serve",
+    "build:start": "cd dist && PORT=3002 npx serve",
     "start": "webpack serve --open --mode development",
     "start:live": "webpack serve --open --mode development --live-reload --hot"
   },
@@ -34,13 +34,10 @@
     "webpack-dev-server": "^4.3.1"
   },
   "dependencies": {
-    "axios": "^0.27.2",
     "bootstrap": "^5.1.3",
-    "dotenv-webpack": "^7.1.0",
     "react": "^17.0.2",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.3.0",
-    "rxjs": "^7.5.5"
+    "react-router-dom": "^6.3.0"
   }
 }
