+++ component-shell/src/App.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import ReactDOM from "react-dom";
+
+import "./index.css";
+
+const App = () => (
+  <div className="container">
+    <div>Name: component-shell</div>
+    <div>Framework: react</div>
+    <div>Language: TypeScript</div>
+    <div>CSS: Empty CSS</div>
+  </div>
+);
+ReactDOM.render(<App />, document.querySelector("#app"));
