+++ component-shell/src/components/home/Home.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <div>
+      <h1 className="text-center text-white">Micro frontend boilerplate</h1>
+    </div>
+  );
+}
