+++ component-shell/src/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>component-shell</title>
+</head>
+
+<body>
+  <div id="app"></div>
+</body>
+
+</html>
