+++ component-shell/webpack.config.js
@@ -4,7 +4,7 @@ const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPl
 const deps = require("./package.json").dependencies;
 module.exports = {
   output: {
-    publicPath: "http://localhost:3000/",
+    publicPath: "http://localhost:3002/",
   },
 
   resolve: {
@@ -12,7 +12,7 @@ module.exports = {
   },
 
   devServer: {
-    port: 3000,
+    port: 3002,
     historyApiFallback: true,
   },
 
@@ -41,15 +41,14 @@ module.exports = {
 
   plugins: [
     new ModuleFederationPlugin({
-      name: "application_shell",
+      name: "component_shell",
       filename: "remoteEntry.js",
-      remotes: {
-        application: "application_shell@http://localhost:3000/remoteEntry.js",
-        authentication:
-          "authentication_shell@http://localhost:3001/remoteEntry.js",
-        component: "component_shell@http://localhost:3002/remoteEntry.js",
+      remotes: {},
+      exposes: {
+        "./Header": "./src/components/header/Header.tsx",
+        "./Footer": "./src/components/footer/Footer.tsx",
+        "./Home": "./src/components/home/Home.tsx",
       },
-      exposes: {},
       shared: {
         ...deps,
         react: {
