+++ package.json
@@ -4,6 +4,8 @@
   "description": "install and run  \r ```\r docker-compose up\r ```",
   "main": "index.js",
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.13.0",
+    "@typescript-eslint/parser": "^5.13.0",
     "eslint": "^8.15.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard": "^17.0.0",
@@ -17,7 +19,8 @@
     "eslint-plugin-unicorn": "^42.0.0",
     "husky": "^8.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "typescript": "~4.6.3"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
